Updated mouse control, DQ

+++ DungeonQuest/Drawable.cpp
@@ -20,8 +20,8 @@ void Drawable::draw()
 			p_SOL_SB->begin();
 
 			const glm::vec4 uvRect(0.0f, 0.0f, 1.0f, 1.0f);
-			glm::vec4 destRect((float)m_pos.x, (float)m_pos.y, 
-				               (float)m_width, (float)m_height);
+			glm::vec4 destRect(m_pos.x, m_pos.y, 
+				               m_width, m_height);
 
             p_SOL_SB->draw(destRect, uvRect, m_textureID, 0.0f, m_colour);
