+++ DungeonQuest/UIButton.cpp
@@ -2,7 +2,7 @@
 
 #include <iostream>
 
-UIButton::UIButton(glm::vec2 v, int width, int height, Solengine::SpriteBatch* batch,
+UIButton::UIButton(glm::vec2 v, float width, float height, Solengine::SpriteBatch* batch,
 	               GLuint texture, Solengine::ColourRGBA8 col)
 {
 	m_pos = (glm::ivec2)v;
