+++ DungeonQuest/UIButton.h
@@ -5,7 +5,7 @@
 class UIButton : public UIElement
 {
 public:
-    UIButton(glm::vec2 v, int width, int height, Solengine::SpriteBatch* spriteBatch,
+    UIButton(glm::vec2 v, float width, float height, Solengine::SpriteBatch* spriteBatch,
 		     GLuint texture = -1, 
 		     Solengine::ColourRGBA8 col = { 255, 255, 255, 255 });
 	~UIButton();
